Enhancement: optimize the filepath and keep snapshot of table.
The relnode is the prefix of filepath for manifest file and data file, there are lots of
files with same prefix for one table and this will increase the accessing latency.

To solve this problem, we use last two parts of uuid as prefix (/a8e2/6d188f753e29/)
instead of relnode(/16384/) for data files, this will disperse data files of one table
and take advantage of load balance of remote storage(oss).

Currently, we do not remove any meta files and data files, metadata in catalog is also
not deleted(vacuum is disabled), so that we can undo any operations, such as drop/truncate,
also we can access any snapshot of table.