Fix: refresh shared buffers and disable hint bits.
Catalog is modified on QD, QE can only read catalog into shared buffers from UnionStore.
When read catalog cached in shared buffers on QE, we need to validate the buffer to check
if it is out of date.

If current read LSN is equal or smaller than LSN of buffer, the buffer is valid; otherwise,
the buffer is out of date, we need to re-read the page from UnionStore into shared buffers.

In InitProcessing Mode, the postgres backends have newest snapshot to see all catalog changes before(dirty read), may set hint bits to tuples which are not visible for current snapshot.
This will cause those tuples invisible for backends which should see those tuples, then cause
data corrupted. To avoid this, we do not set hint bits in InitProcessing Mode.