Add new hashdata am oid to RelationIsNonblockRelation
Add a new am oid to RelationIsNonblockRelation, so that the new storage could access some
specific code. The reason for disable update/delete trigger is that the cost of random
access to these storages is too high. But only after trigger will call fetch api, so we
should all before trigger for delete/update.