Fix: gpdiffcheck/query_finish_pending/segspace/dispatch cases
For query_finish_pending case, the result of query 'select i1 from \_tmp_table order by i2 limit 3'
 is mutable with limit and order by clauses and fault injection because we use the randomly
distribution as default for table, so modify the sql statement with count.

For other cases, modify the 'Create Table' statements and add outputs.

For single query transaction, if we are not in transaction block and no xid is assigned, do not
send commit command to QE.