Analyze table change to happen on QD.
When we collect data from segments to QD, this will change the physical order of the data.
Such as in segment 1 the data is 1,3,5,7,9. And in segment 2 the data is 2,4,6,8,10. In each
segment the data is ordered, and correlation is 1 in each segment. But after we collect the data
 to QD, it may be 1,3,5,2,4,7,9,6,8,10. And the correlation is 0.3 or something else and it is
not stable.

So get correlations from QD on cloud version.