Fix: use XLogRecGetBlockTag to get block number in brin_redo.
The redo function is called in UnionStore to replay the WAL on page, but walredo
procedure could replay WAL on most one page every time.

For WALs related with multi pages, some pages may be skipped, do not replay WAL on
those pages. The related buffer with those pages may be invalid, so we use
XLogRecGetBlockTag here to avoid touching the invalid buffer.