Feature: optimize simple aggregate query
Currently, Max/Min stats on columns is supported in PAX. When we write new data,
the stats is stored in manifest. But if we delete/update the tuples, the stats will
not be updated, that means the stats is not accurate to answer the query, can only be
used as Max/Min filters.

If we rewrite all the data after delete/update, the stats is re-collected and accurate.
So we add new write policy 'COPY ON WRITE' for hashdata table: create table test
(c1 int, c2 bigint) with (write_policy='cow');

For hashdata table with COPY ON WRITE write policy, the Max/Min stats is accurate, we
use this stats as direct results of simple aggregate max/min on single column, such as
query select max(c1), min(c2) from test;

We use Custom Scan on manifest to get the results instead of Aggregate Node with SeqScan
to accelerate such query.

For more details, see https://hashdata.feishu.cn/docx/ExmXd5clOoUJJyxohMOca6qYnlb