Collect catalog and dispatch to qes
Dispatch the catalogs that need to be accessed from QD to QE, and no longer directly access system tables on QE. This will reduce the access pressure on the Union Store and the visibility issues caused by the inconsistency of the catalogs seen on QD and QE.

QE receives the Plan from QD and converts it to PlanState by the ExecutorStart function. In this process, the Catalog needs to be read. QD will perform the same process to convert the Plan to PlanState. The ExecutorStart in QD and QE execute the same code and access the same Catalog. We can collect the Catalogs accessed by the ExecutorStart in QD and distribute them to QE.

QD accesses the Catalog through 4 functions. SearchSysCache SearchSysCacheList systable_getnext systable_getnext_ordered. We embed code in these functions to collect Catalog tuples and distribute them to QE. QE uses hashtable and list to store the distributed data and implements a simple am. systable_getnext systable_getnext_ordered calls this simple am to access the catalog.