Support Subquery in Append Agg.
For a append agg inside a subquery, we will use its target list to match
materialized view.
However, Postgres will remove unused columns of subquery that do not
exit in upper query in a hacky way: make NUll for target entries.
It will make us fail to match view and rewrite as we only support
exactly match for now.

Workaround for this with GUC if we are allowed to  attmpt to answer query.