Fix: auto analyze inherited table error
It will coredump when auto analyze the inherited table, this reason is that it will enter
the acquire_sample_rows_dispatcher method:

if (Gp_role == GP_ROLE_DISPATCH && ENABLE_DISPATCH())
{
        return acquire_sample_rows_dispatcher(onerel,
                                            true, /* inherited stats */
                                            elevel,
                                            rows,
                                            targrows,
                                            totalrows,
                                            totaldeadrows);
}

but if we analyze the table manually, we hold dispatch in hashdata_ProcessUtility.
so we will follow the manual analyze logic, just igore this routine in serverlese mode.