Force CTAS distribution to random in serverless mode.
For CRATE..AS, taregt table's distribution policy could be derived
from the Query of AS part.

create materialized view cloud_ctas_mv as select a, count(b) from
cloud_ctas_t0 group by a with no data;

The locus of cloud_ctas_mv may be Hashed by cloumn a as it's a agg
with group by.
It's ok for CBDB, but in serverless mode, the underlying data is
random, we can't store a distribution policy with distkeys or
numsegments for that.Else, will get error when we switch to a
cluster with more or less segments.

We have done something in extensions, but it didn't take effect due
to the architecture and process of utility hooks.

This is the last resort in core codes.

Authored-by: Zhang Mingli avamingli@gmail.com