Eliminate const expressions for Append AGG.
Const expressions like where 1 = 1 and a > 1 will be processed
to where a > 1 by planner.
Quals like: 1 = 1 is always TRUE, for a AND expression that's
useless.

But as we store MV's view query as it was originally, the parse
tree processed by planner may not match MV's exactly.
Process that quals during Append AGG to fix.

Authored-by: Zhang Mingli avamingli@gmail.com